Update subdirectory structure; move case-study module.
