+++ src/components/indices/EntrepreneurshipIndex.jsx
@@ -1,0 +1,20 @@
+// React 
+import React from 'react'
+
+// React Semantic
+import {
+  Label,
+  Grid,
+  Card,
+  Icon,
+  Image,
+  Button,
+} from 'semantic-ui-react'
+
+export default function EntrepreneurshipIndex(props) {
+  return (
+    <>
+
+    </>
+  )
+}
